requirements.txt
djangorestframework==3.7.3
djangorestframework-jwt 
django-cors-headers
psycopg2 == 2.7.3
Pillow == 4.3.0
Django==1.11.7
django-cors-headers
requests